tower: plumb interrupt numbers through the AADL/STM32 backends because eChronos needs them

+++ tower-aadl/src/Tower/AADL/AST/Common.hs
@@ -94,7 +94,7 @@ endPointId :: Endpoint -> ChanId
 endPointId ep = case ep of
   InputEp  rx -> inputId  rx
   OutputEp tx -> outputId tx
-  SignalEp s  -> signalNumber s
+  SignalEp s  -> fromIntegral (signalNumber s)
 
 newChan :: LocalId -> Endpoint -> ThdIds
 newChan l ep =
